refactor : step3 요구사항 추가 반영
